More minor documentation fixes.
